services/notification-worker/requirements.txt
nats-py>=2.7
fastapi>=0.111
uvicorn[standard]>=0.29
prometheus-client>=0.20
